refactor: reusable squads list
