+++ packages/shared/src/components/widgets/SocialShare.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement, useContext, useMemo } from 'react';
+import React, { ReactElement, useContext } from 'react';
 import {
   getEmailShareLink,
   getFacebookShareLink,
@@ -9,7 +9,7 @@ import {
   getWhatsappShareLink,
   ShareProvider,
 } from '../../lib/share';
-import { ShareText, SocialShareIcon } from './SocialShareIcon';
+import { SocialShareIcon } from './SocialShareIcon';
 import { Post } from '../../graphql/posts';
 import MailIcon from '../icons/Mail';
 import TwitterIcon from '../icons/Twitter';
@@ -22,13 +22,7 @@ import { FeedItemPosition, postAnalyticsEvent } from '../../lib/feed';
 import { AnalyticsEvent, Origin } from '../../lib/analytics';
 import AnalyticsContext from '../../contexts/AnalyticsContext';
 import { Comment, getCommentHash } from '../../graphql/comments';
-import { useAuthContext } from '../../contexts/AuthContext';
-import { SourcePermissions } from '../../graphql/sources';
-import SourceProfilePicture from '../profile/SourceProfilePicture';
-import { verifyPermission } from '../../graphql/squads';
-import { useCreateSquadModal } from '../../hooks/useCreateSquadModal';
 import FeaturesContext from '../../contexts/FeaturesContext';
-import PlusIcon from '../icons/Plus';
 import { IconSize } from '../Icon';
 import { useSharePost } from '../../hooks/useSharePost';
 import MenuIcon from '../icons/Menu';
@@ -36,6 +30,7 @@ import CopyIcon from '../icons/Copy';
 import { usePostToSquad } from '../../hooks';
 import { SocialShareContainer } from './SocialShareContainer';
 import { useCopyLink } from '../../hooks/useCopyLink';
+import { SquadsToShare } from '../squads/SquadsToShare';
 
 interface SocialShareProps {
   origin: Origin;
@@ -59,18 +54,12 @@ export const SocialShare = ({
   const href = isComment
     ? `${post?.commentsPermalink}${getCommentHash(comment.id)}`
     : post?.commentsPermalink;
-  const { squads } = useAuthContext();
   const [copying, copyLink] = useCopyLink(() => href);
   const link = isComment
     ? `${post?.commentsPermalink}${getCommentHash(comment.id)}`
     : post?.commentsPermalink;
   const { trackEvent } = useContext(AnalyticsContext);
-  const { hasSquadAccess, isFlagsFetched } = useContext(FeaturesContext);
-  const { openNewSquadModal } = useCreateSquadModal({
-    hasSquads: !!squads?.length,
-    hasAccess: hasSquadAccess,
-    isFlagsFetched,
-  });
+  const { hasSquadAccess } = useContext(FeaturesContext);
   const { openNativeSharePost } = useSharePost(Origin.Share);
   const trackClick = (provider: ShareProvider) =>
     trackEvent(
@@ -95,49 +84,14 @@ export const SocialShare = ({
     trackClick(ShareProvider.CopyLink);
   };
 
-  const list = useMemo(
-    () =>
-      squads
-        ?.filter(
-          (squadItem) =>
-            squadItem.active &&
-            verifyPermission(squadItem, SourcePermissions.Post),
-        )
-        .map((squad) => (
-          <button
-            type="button"
-            className="flex overflow-hidden flex-col items-center w-16 text-center"
-            key={squad.id}
-            onClick={(e) => onSubmitPost(e, squad.id, commentary)}
-            disabled={isPosting}
-          >
-            <SourceProfilePicture source={squad} />
-            <ShareText className="mt-2 max-w-full truncate">
-              @{squad.handle}
-            </ShareText>
-          </button>
-        )),
-    [squads, onSubmitPost, commentary, isPosting],
-  );
-
   return (
     <>
       {hasSquadAccess && !isComment && !post.private && (
         <SocialShareContainer title="Share with your squad">
-          {list ?? (
-            <SocialShareIcon
-              onClick={() =>
-                openNewSquadModal({
-                  origin: Origin.Share,
-                  redirectAfterCreate: false,
-                })
-              }
-              pressed={copying}
-              icon={<PlusIcon className="text-theme-label-invert" />}
-              className="!rounded-full btn-primary-cabbage"
-              label="New Squad"
-            />
-          )}
+          <SquadsToShare
+            onClick={(e, squad) => onSubmitPost(e, squad.id, commentary)}
+            isLoading={isPosting}
+          />
         </SocialShareContainer>
       )}
       <SocialShareContainer title="Share externally" className="mt-4">
